requirements.txt
Django==5.2.6
python-dotenv==1.0.0
openpyxl==3.1.5
Pillow==12.0.0
psycopg2-binary==2.9.9
mysqlclient==2.2.4
